move TestTool from shared to settings
