+++ frontend/src/modules/settings/tabs/TestCalibrationTab.tsx
@@ -5,7 +5,7 @@
  */
 import React from 'react';
 import { Grid, makeStyles, Theme } from '@material-ui/core';
-import TestTool from '../../shared/TestTool';
+import TestTool from './TestTool';
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
