+++ frontend/src/modules/settings/tabs/index.ts
@@ -8,3 +8,4 @@
 export { default as DisplayTab } from './DisplayTab';
 export { default as InfoTab } from './InfoTab';
 export { default as TestCalibrationTab } from './TestCalibrationTab';
+export { default as TestTool } from './TestTool';
