+++ frontend/src/modules/shared/index.ts
@@ -1,3 +1,2 @@
 export * from './ModalPopup';
 export { default as TabPanel } from './TabPanel';
-export { default as TestTool } from './TestTool';
